Beta Commit
Has the basic functions and types... Beta testing needs to be done.
Default function arguments need to be given and convenience functions
need to be made.
Functions still need to be noted as discardable or to be discarded when
used.
Object types for various should be made to reflect the C++ library it
was wrapped from.
Garbage collection safety still needs to be guaranteed, but it is safe
to use if memory was copied to the C side.